Can some decorators make our life easier? Let's find out!
The idea is to implement some decorators for the handlers that check,
based on the arguments, if the request is valid or not.
For example many requests can be made only if the contest is running.
For checking this constraint we need to basically copy and paste the
same piece of code to check if the user has some time left and raise
usually the same error message.

The problem with the committed approach is that we need to change a lot
of the code to adapt to this strategy (the arguments have to have a
descriptive name to help the guesser).

At this stage the implemented method adds an overhead because it needs
to query the database for many things (sometimes even multiple times).
Without a cache it may cause an issue (needs benchmarking!). Personally
I think that this problem will occur only if we decide to scale for
bigger contests, we'll see...