Contests can have a per-user window shorter than the contest duration
In USACO-like competitions the contests are usually very long (days) but
each contestant can compete for a short (hours) time window. In this case
the window is started at the login and will end after "window_duration"
or at the contest end (whichever comes first).

In the DB the delay from the start of the contest is stored in the users
table. The extra time of the contest won't extend the window but the
user extra time will do.